"docker-compile" --> "docker compile"
Replacing all instances of "docker-compile" with "docker compile" for compatibility with v2.* of Docker Compile.
v2 breaks the code in hydra.  This simple change will ensure it is compatible with v2. Docker Compile v1 is losing support at the end of June so it's best that we change this now and enforce users to make sure they are on the correct version of Docker Compile.

* This change would not be backwards compatible with v1. *