set_docker_compose function created
Detects and sets the correct syntax for docker compose. If v1 is the only version found it can use it but it will recommend v2+ if it is not found. If neither version is found it stops the script and lets the user know they will need to install it.

This new code only checks during the relevant functions it needs to set the syntax.