.env file updates during hydra install
This code will allow the user to input custom data that will update the .env file if it detects the default information is being used. It will also check if the GITHUB_TOKEN is accessible. If it is not it will request the user to send the correct github token.

This code is useful to simplify setting up the sdk using custom p12 files.  

This does not change the ability for developers to manually edit the .env file themselves.  This only asks for details from the user if the data is set as the default variables. Or if the github token is invalid.

In another pull request I have sent it will also use the custom p12 data in the .env to compile the custom p12 files during the build process.