No spaces allowed
When setting the $PROJECT_NAME I noticed spaces and uppercase letters would cause issues in the build process.  So I added a loop that prevents the user from adding spaces and even if they add uppercase letters, the code will change it to lowercase.

For each p12 filename, alias and password no spaces are allowed and a loop has been added to prevent this.  These all seem to be okay with uppercase or lowercase letters.